- create an inspector service class to enable tdd of functions and reusability
